requirements.txt
# Core dependencies for OSINT Pipeline
# Social Media APIs
tweepy==4.16.0
praw==7.8.1
prawcore==2.4.0
facebook-sdk==3.1.0
discord.py==2.6.3
instaloader==4.14.2
PyGithub==2.8.1
linkedin-api==2.3.1
snscrape==0.7.0.20230622

# Web scraping and HTTP requests
requests==2.32.5
beautifulsoup4==4.13.5
lxml==5.4.0
urllib3==2.5.0
aiohttp==3.12.15
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
multidict==6.6.4
yarl==1.20.1
frozenlist==1.7.0
propcache==0.3.2

# Data processing and analysis
pandas==2.3.2
numpy==2.3.3
scipy==1.16.2
matplotlib==3.10.6
scikit-learn==1.7.2
joblib==1.5.2
threadpoolctl==3.6.0

# Text processing and NLP
textblob==0.19.0
langdetect==1.0.9
nltk==3.9.1
regex==2025.9.18

# Database and data storage
sqlalchemy==2.0.43
records==0.6.0
tablib==3.8.0
openpyxl==3.1.5
et-xmlfile==2.0.0

# Environment and configuration management
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Authentication and security
PyJWT==2.10.1
cryptography==46.0.1
cffi==2.0.0
pycparser==2.23
PyNaCl==1.6.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
PySocks==1.7.1

# Utilities and tools
click==8.3.0
tqdm==4.67.1
schedule==1.2.2
update-checker==0.18.0
docopt==0.6.2
attrs==25.3.0
packaging==25.0
setuptools==80.9.0
wheel==0.45.1
pip==25.2

# GUI and visualization support
pillow==11.3.0
fonttools==4.60.0
kiwisolver==1.4.9
cycler==0.12.1
contourpy==1.3.3

# Network and websocket support
websocket-client==1.8.0

# Additional utilities
six==1.17.0
pyparsing==3.2.5
soupsieve==2.8
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
typing-extensions==4.15.0
filelock==3.19.1
greenlet==3.2.4
colorama==0.4.6
telethon==1.29.1
mastodon.py==1.6.0